DOMA-2815 move recipient field from domain common to domain acquiring
