+++ apps/condo/domains/acquiring/schema/AcquiringIntegrationContext.js
@@ -5,7 +5,8 @@ const { Text, Relationship } = require('@keystonejs/fields')
 const { Json } = require('@core/keystone/fields')
 const { GQLListSchema } = require('@core/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
-const { SENDER_FIELD, DV_FIELD, RECIPIENT_FIELD } = require('@condo/domains/common/schema/fields')
+const { SENDER_FIELD, DV_FIELD } = require('@condo/domains/common/schema/fields')
+const { RECIPIENT_FIELD } = require('@condo/domains/acquiring/schema/fields/Recipient.js')
 const { ACQUIRING_INTEGRATION_FIELD } = require('@condo/domains/acquiring/schema/fields/relations')
 const access = require('@condo/domains/acquiring/access/AcquiringIntegrationContext')
 const { hasValidJsonStructure, hasDvAndSenderFields } = require('@condo/domains/common/utils/validation.utils')
