+++ apps/condo/domains/billing/schema/BillingReceipt.js
@@ -2,15 +2,15 @@
  * Generated by `createschema billing.BillingReceipt 'context:Relationship:BillingIntegrationOrganizationContext:CASCADE; importId?:Text; property:Relationship:BillingProperty:CASCADE; account:Relationship:BillingAccount:CASCADE; period:CalendarDay; raw:Json; toPay:Text; services:Json; meta:Json'`
  */
 
+const access = require('@condo/domains/billing/access/BillingReceipt')
+
 const { get } = require('lodash')
 const { Text, Relationship } = require('@keystonejs/fields')
 
 const { GQLListSchema } = require('@core/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
 
-const { SENDER_FIELD, DV_FIELD } = require('@condo/domains/common/schema/fields')
-const access = require('@condo/domains/billing/access/BillingReceipt')
-const { MONEY_AMOUNT_FIELD } = require('@condo/domains/common/schema/fields')
+const { SENDER_FIELD, DV_FIELD, MONEY_AMOUNT_FIELD } = require('@condo/domains/common/schema/fields')
 const { hasDvAndSenderFields } = require('@condo/domains/common/utils/validation.utils')
 const { DV_UNKNOWN_VERSION_ERROR, WRONG_TEXT_FORMAT } = require('@condo/domains/common/constants/errors')
