+++ apps/condo/domains/billing/schema/fields/BillingReceipt/Recipient.js
@@ -1,6 +1,6 @@
 const { render } = require('@condo/domains/billing/schema/fields/utils/json.utils')
 const { BILLING_RECEIPT_RECIPIENT_FIELD_NAME, BILLING_RECEIPT_RECIPIENT_INPUT_NAME } = require('@condo/domains/billing/constants/constants')
-const { RECIPIENT_FIELDS_DEFINITION, RECIPIENT_FIELD: COMMON_RECIPIENT_FIELD, RECIPIENT_QUERY_LIST } = require('@condo/domains/common/schema/fields.js')
+const { RECIPIENT_FIELDS_DEFINITION, RECIPIENT_FIELD: COMMON_RECIPIENT_FIELD, RECIPIENT_QUERY_LIST } = require('@condo/domains/acquiring/schema/fields/Recipient.js')
 
 
 const RECIPIENT_GRAPHQL_TYPES = `
