+++ apps/condo/domains/common/schema/fields.js
@@ -1,4 +1,3 @@
-const Ajv = require('ajv')
 const { Relationship, Select, Integer, Text } = require('@keystonejs/fields')
 const { Decimal } = require('@keystonejs/fields')
 const { Json, SignedDecimal } = require('@core/keystone/fields')
@@ -7,7 +6,6 @@ const { normalizePhone } = require('@condo/domains/common/utils/phone')
 const { hasValidJsonStructure } = require('@condo/domains/common/utils/validation.utils')
 const { JSON_UNKNOWN_VERSION_ERROR, REQUIRED_NO_VALUE_ERROR, JSON_EXPECT_OBJECT_ERROR } = require('@condo/domains/common/constants/errors')
 const { ADDRESS_META_FIELD_GRAPHQL_TYPES, ADDRESS_META_SUBFIELDS_QUERY_LIST } = require('@condo/domains/property/schema/fields/AddressMetaField')
-const { render, getValidator } = require('@condo/domains/common/schema/json.utils.js')
 const { ISO_CODES } = require('../constants/currencies')
 const { UNIT_TYPES, FLAT_UNIT_TYPE } = require('@condo/domains/property/constants/common')
 
@@ -171,55 +169,6 @@ const UNIT_TYPE_FIELD = {
     defaultValue: FLAT_UNIT_TYPE,
 }
 
-
-const RECIPIENT_FIELDS_DEFINITION = {
-    name: 'String',
-    bankName: 'String',
-    territoryCode: 'String',
-    offsettingAccount: 'String',
-    tin: 'String!',
-    iec: 'String!',
-    bic: 'String!',
-    bankAccount: 'String!',
-}
-
-const RECIPIENT_FIELD_NAME = 'RecipientField'
-const RECIPIENT_INPUT_NAME = 'RecipientFieldInput'
-
-const ajv = new Ajv()
-
-const RecipientSchema = {
-    type: 'object',
-    properties: Object.assign({},
-        ...Object.keys(RECIPIENT_FIELDS_DEFINITION).map((field) => ({ [field]: { type: 'string' } })),
-    ),
-    required: Object.keys(RECIPIENT_FIELDS_DEFINITION).filter(fieldName => RECIPIENT_FIELDS_DEFINITION[fieldName].slice(-1) === '!'),
-    additionalProperties: false,
-}
-
-const RECIPIENT_GRAPHQL_TYPES = `
-    type ${RECIPIENT_FIELD_NAME} {
-        ${render(RECIPIENT_FIELDS_DEFINITION)}
-    }
-    
-    input ${RECIPIENT_INPUT_NAME} {
-        ${render(RECIPIENT_FIELDS_DEFINITION)}
-    }
-`
-
-const RECIPIENT_FIELD = {
-    schemaDoc: 'Recipient. Should contain all meta information to identify the organization',
-    type: Json,
-    isRequired: true,
-    graphQLAdminFragment: `{ ${Object.keys(RECIPIENT_FIELDS_DEFINITION).join(' ')} }`,
-    extendGraphQLTypes: [RECIPIENT_GRAPHQL_TYPES],
-    graphQLReturnType: RECIPIENT_FIELD_NAME,
-    graphQLInputType: RECIPIENT_INPUT_NAME,
-    hooks: {
-        validateInput: getValidator(ajv.compile(RecipientSchema)),
-    },
-}
-
 module.exports = {
     DV_FIELD,
     SENDER_FIELD,
@@ -236,6 +185,4 @@ module.exports = {
     NON_NEGATIVE_MONEY_FIELD,
     IMPORT_ID_FIELD,
     UNIT_TYPE_FIELD,
-    RECIPIENT_FIELD,
-    RECIPIENT_FIELDS_DEFINITION,
 }
